[refactor/#84] 패키지 구조 이동, 시야 mockdata 삭제
